+++ include/seqan/arg_parse.h
@@ -52,6 +52,10 @@
 #include <unistd.h>
 #endif
 
+#include <regex>
+#include <vector>
+#include <algorithm>
+
 // ===========================================================================
 // The ArgParse Headers.
 // ===========================================================================
@@ -70,5 +74,6 @@
 #include <seqan/arg_parse/arg_parse_ctd_support.h>
 #include <seqan/arg_parse/arg_parse_parse.h>
 
+#include <seqan/arg_parse/arg_parse_version_check.h>
 
 #endif  // SEQAN_INCLUDE_SEQAN_ARG_PARSE_H_