+++ include/seqan/stream/formatted_file.h
@@ -716,6 +716,26 @@ inline bool open(FormattedFile<TFileFormat, TDirection, TSpec> & file,
 }
 
 // ----------------------------------------------------------------------------
+// Function isOpen()
+// ----------------------------------------------------------------------------
+
+/*!
+ * @fn FormattedFile#isOpen
+ * @brief Determines whether a FormattedFile is currently open.
+ *
+ * @signature bool isOpen(file);
+ *
+ * @param[in]     file The FormattedFile to check.
+ * @return bool <tt>true</tt> if file is currently open, <tt>false</tt> otherwise.
+ */
+
+template <typename TFileFormat, typename TDirection, typename TSpec>
+inline bool isOpen(FormattedFile<TFileFormat, TDirection, TSpec> const & file)
+{
+    return file.stream.file.is_open();
+}
+
+// ----------------------------------------------------------------------------
 // Function close()
 // ----------------------------------------------------------------------------